Laravel conventions say: Pluralise table names
See: https://laravel.com/docs/11.x/eloquent#table-names

Under the assumption that we're creating this database from scratch (`php artisan migrate:fresh`) we can retroactively change the name of the table in the migration.

If this was something that was already running in production, a regular migration changing the name would be more appropriate (although I would argue that if this is already running in production, the value of changing this is dubious).