Add custom PAT to .env for convenience
This is a handy way of sharing .env files across users without necessarily committing it to Git (although we'll do it anyway).

In a real scenario, we wouldn't commit the .env files.

This still requires the right PAT to be seeded into the database, and for it to match the entry in the .env file.